Fixed admin
# Conflicts:
#	.gitignore

+++ src/main/java/Controller/Meeting_RoomRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface Meeting_RoomRepository extends CrudRepository<Meeting_Room, Long> {
+public interface Meeting_RoomRepository extends CrudRepository<Meeting_Room, Integer> {
 
 }
